LPS-169327 Merge EditClientExtensionEntryPartDisplayContext into EditClientExtensionEntryDisplayContext
