--- modules/apps/client-extension/client-extension-web/src/main/java/com/liferay/client/extension/web/internal/display/context/EditClientExtensionEntryPartDisplayContext.java
@@ -1,144 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.client.extension.web.internal.display.context;
-
-import com.liferay.client.extension.model.ClientExtensionEntry;
-import com.liferay.client.extension.type.CET;
-import com.liferay.frontend.taglib.clay.servlet.taglib.util.SelectOption;
-import com.liferay.petra.string.CharPool;
-import com.liferay.petra.string.StringPool;
-import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.language.LanguageUtil;
-import com.liferay.portal.kernel.model.PortletCategory;
-import com.liferay.portal.kernel.theme.ThemeDisplay;
-import com.liferay.portal.kernel.util.ListUtil;
-import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portal.kernel.util.Validator;
-import com.liferay.portal.kernel.util.WebKeys;
-import com.liferay.portal.util.WebAppPool;
-
-import java.util.ArrayList;
-import java.util.Comparator;
-import java.util.List;
-import java.util.Objects;
-
-import javax.portlet.PortletRequest;
-
-/**
- * @author Iván Zaera Avellón
- */
-public class EditClientExtensionEntryPartDisplayContext<T extends CET> {
-
-	public EditClientExtensionEntryPartDisplayContext(
-			T cet, ClientExtensionEntry clientExtensionEntry,
-			PortletRequest portletRequest)
-		throws PortalException {
-
-		_cet = cet;
-		_clientExtensionEntry = clientExtensionEntry;
-		_portletRequest = portletRequest;
-	}
-
-	public T getCET() {
-		return _cet;
-	}
-
-	public List<SelectOption> getPortletCategoryNameSelectOptions(
-		String selectedPortletCategoryName) {
-
-		if (Validator.isBlank(selectedPortletCategoryName)) {
-			selectedPortletCategoryName = "category.remote-apps";
-		}
-
-		List<SelectOption> selectOptions = new ArrayList<>();
-
-		boolean found = false;
-
-		ThemeDisplay themeDisplay = (ThemeDisplay)_portletRequest.getAttribute(
-			WebKeys.THEME_DISPLAY);
-
-		PortletCategory rootPortletCategory = (PortletCategory)WebAppPool.get(
-			themeDisplay.getCompanyId(), WebKeys.PORTLET_CATEGORY);
-
-		for (PortletCategory portletCategory :
-				rootPortletCategory.getCategories()) {
-
-			selectOptions.add(
-				new SelectOption(
-					LanguageUtil.get(
-						themeDisplay.getLocale(), portletCategory.getName()),
-					portletCategory.getName(),
-					selectedPortletCategoryName.equals(
-						portletCategory.getName())));
-
-			if (Objects.equals(
-					portletCategory.getName(), "category.remote-apps")) {
-
-				found = true;
-			}
-		}
-
-		if (!found) {
-			selectOptions.add(
-				new SelectOption(
-					LanguageUtil.get(
-						themeDisplay.getLocale(), "category.remote-apps"),
-					"category.remote-apps",
-					Objects.equals(
-						selectedPortletCategoryName, "category.remote-apps")));
-		}
-
-		return ListUtil.sort(
-			selectOptions,
-			new Comparator<SelectOption>() {
-
-				@Override
-				public int compare(
-					SelectOption selectOption1, SelectOption selectOption2) {
-
-					String label1 = selectOption1.getLabel();
-					String label2 = selectOption2.getLabel();
-
-					return label1.compareTo(label2);
-				}
-
-			});
-	}
-
-	public String[] getStrings(String urls) {
-		String[] strings = StringUtil.split(urls, CharPool.NEW_LINE);
-
-		if (strings.length == 0) {
-			return _EMPTY_STRINGS;
-		}
-
-		return strings;
-	}
-
-	public boolean isNew() {
-		if (_clientExtensionEntry == null) {
-			return true;
-		}
-
-		return false;
-	}
-
-	private static final String[] _EMPTY_STRINGS = {StringPool.BLANK};
-
-	private final T _cet;
-	private final ClientExtensionEntry _clientExtensionEntry;
-	private final PortletRequest _portletRequest;
-
-}
