+++ modules/apps/client-extension/client-extension-web/src/main/java/com/liferay/client/extension/web/internal/portlet/action/EditClientExtensionEntryMVCRenderCommand.java
@@ -21,7 +21,6 @@ import com.liferay.client.extension.type.factory.CETFactory;
 import com.liferay.client.extension.web.internal.constants.ClientExtensionAdminPortletKeys;
 import com.liferay.client.extension.web.internal.constants.ClientExtensionAdminWebKeys;
 import com.liferay.client.extension.web.internal.display.context.EditClientExtensionEntryDisplayContext;
-import com.liferay.client.extension.web.internal.display.context.EditClientExtensionEntryPartDisplayContext;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCRenderCommand;
 import com.liferay.portal.kernel.util.ParamUtil;
@@ -71,11 +70,6 @@ public class EditClientExtensionEntryMVCRenderCommand
 				ClientExtensionAdminWebKeys.
 					EDIT_CLIENT_EXTENSION_ENTRY_DISPLAY_CONTEXT,
 				new EditClientExtensionEntryDisplayContext(
-					cet, clientExtensionEntry, renderRequest));
-			renderRequest.setAttribute(
-				ClientExtensionAdminWebKeys.
-					EDIT_CLIENT_EXTENSION_ENTRY_PART_DISPLAY_CONTEXT,
-				new EditClientExtensionEntryPartDisplayContext(
 					cet, clientExtensionEntry, renderRequest));
 
 			return "/admin/edit_client_extension_entry.jsp";
