+++ modules/apps/client-extension/client-extension-web/src/main/resources/META-INF/resources/admin/edit_theme_favicon.jsp
@@ -17,9 +17,9 @@
 <%@ include file="/admin/init.jsp" %>
 
 <%
-EditClientExtensionEntryPartDisplayContext<ThemeFaviconCET> editClientExtensionEntryPartDisplayContext = (EditClientExtensionEntryPartDisplayContext)renderRequest.getAttribute(ClientExtensionAdminWebKeys.EDIT_CLIENT_EXTENSION_ENTRY_PART_DISPLAY_CONTEXT);
+EditClientExtensionEntryDisplayContext<ThemeFaviconCET> editClientExtensionEntryDisplayContext = (EditClientExtensionEntryDisplayContext)renderRequest.getAttribute(ClientExtensionAdminWebKeys.EDIT_CLIENT_EXTENSION_ENTRY_DISPLAY_CONTEXT);
 
-ThemeFaviconCET themeFaviconCET = editClientExtensionEntryPartDisplayContext.getCET();
+ThemeFaviconCET themeFaviconCET = editClientExtensionEntryDisplayContext.getCET();
 %>
 
 <aui:input label="url" name="url" type="text" value="<%= themeFaviconCET.getURL() %>" />
