+++ modules/apps/client-extension/client-extension-web/src/main/resources/META-INF/resources/admin/edit_theme_js.jsp
@@ -17,9 +17,9 @@
 <%@ include file="/admin/init.jsp" %>
 
 <%
-EditClientExtensionEntryPartDisplayContext<ThemeJSCET> editClientExtensionEntryPartDisplayContext = (EditClientExtensionEntryPartDisplayContext)renderRequest.getAttribute(ClientExtensionAdminWebKeys.EDIT_CLIENT_EXTENSION_ENTRY_PART_DISPLAY_CONTEXT);
+EditClientExtensionEntryDisplayContext<ThemeJSCET> editClientExtensionEntryDisplayContext = (EditClientExtensionEntryDisplayContext)renderRequest.getAttribute(ClientExtensionAdminWebKeys.EDIT_CLIENT_EXTENSION_ENTRY_DISPLAY_CONTEXT);
 
-ThemeJSCET themeJSCET = editClientExtensionEntryPartDisplayContext.getCET();
+ThemeJSCET themeJSCET = editClientExtensionEntryDisplayContext.getCET();
 %>
 
 <aui:input ignoreRequestValue="<%= true %>" label="url" name="url" type="text" value="<%= themeJSCET.getURL() %>" />
